feat: type 각 페이지 폴더로 분리
