+++ footlog/src/components/home/RecommendContainer.tsx
@@ -1,5 +1,5 @@
 import CoursesSlider from '@components/common/CoursesSlider/CoursesSlider';
-import { RecommendCoursesDataTypes } from 'types/CommonTypes';
+import { RecommendCoursesDataTypes } from 'types/common/CommonTypes';
 
 interface RecommendContainerProps {
   title: string;
