+++ footlog/src/core/recommendCoursesData.ts
@@ -1,5 +1,5 @@
 import courseExampleImg from '@public/courseExample.png';
-import { RecommendCoursesDataTypes } from 'types/CommonTypes';
+import { RecommendCoursesDataTypes } from 'types/common/CommonTypes';
 
 export const recommendCoursesData: RecommendCoursesDataTypes[] = [
   { id: 1, imgSrc: courseExampleImg, title: '문경새재 도립공원', subtitle: '경상북도 문경시', isSaved: true },
