+++ footlog/src/types/onboarding/OnboardingTypes.ts
@@ -1,0 +1,5 @@
+export interface OnboardingIconsDataTypes {
+  keyword: string;
+  defaultIcon: React.FC;
+  strokeIcon: React.FC;
+}
